requirements.txt
Flask
docker-py
mock